Refactor server.py to include stop and start server functionality

+++ server.py
@@ -7,9 +7,13 @@ import sys
 def run_server():
     return subprocess.Popen(['waitress-serve', '--port=5000', '--threads=4', 'main:app'], stdout=subprocess.PIPE)
 
+def stop_server(server_process):
+    if server_process:
+        server_process.terminate()
+        server_process.wait()
+
 def restart_server(server_process):
-    server_process.terminate()
-    server_process.wait()
+    stop_server(server_process)
     changes = pull_from_git()
     if 'server.py' in changes:
         os.execv(__file__, sys.argv)
@@ -33,6 +37,14 @@ while True:
     command = input()
     if command == 'restart':
         server_process = restart_server(server_process)
+    elif command == 'stop':
+        stop_server(server_process)
+        server_process = None
+    elif command == 'start':
+        if server_process is None:
+            server_process = run_server()
+        else:
+            print("Server is already running.")
     elif command.startswith('open '):
         file_type = command.split(' ')[1]
         open_newest_file(file_type)
